% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_score.R
\name{add_title_dp}
\alias{add_title_dp}
\title{Add a centered title to a ggplot object}
\usage{
add_title_dp(plot, title_text)
}
\arguments{
\item{plot}{A ggplot object.}

\item{title_text}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Internal plotting helper used in score-related plotting functions.
}
\keyword{internal}
