% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult data example}
\format{
32,651 by 5 numeric matrix
}
\source{
Prepared for examples in the dppca package.
(\emph{We need to add a reference of this dataset})
}
\usage{
data(adult)
}
\description{
A sample dataset used for examples in the dppca package.
}
\keyword{datasets}
