% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lambda.R
\name{compute_np_pca_basis}
\alias{compute_np_pca_basis}
\title{Compute non-private PCA basis}
\usage{
compute_np_pca_basis(X_proc, k)
}
\arguments{
\item{X_proc}{Preprocessed numeric matrix.}

\item{k}{Number of leading components.}
}
\value{
A d x k orthonormal matrix.
}
\description{
Compute non-private PCA basis
}
\keyword{internal}
