% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_score_app.R
\name{dp_score_app}
\alias{dp_score_app}
\title{Run DP PCA score-plot Shiny app for user-supplied data}
\usage{
dp_score_app(X, G = NULL, title = "DP PCA Score Plot (User data)")
}
\arguments{
\item{X}{Numeric matrix/data.frame (n x p).}

\item{G}{Optional group vector (length n) or a column name in X.}

\item{title}{App title.}
}
\description{
Run DP PCA score-plot Shiny app for user-supplied data
}
