% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eur_map}
\alias{eur_map}
\title{Europe map PCA data}
\format{
A data frame or matrix.
}
\source{
Prepared for examples in the dppca package.
(\emph{We need to add a reference of this dataset})
}
\usage{
data(eur_map)
}
\description{
A dataset used for PCA score plot examples.
}
\keyword{datasets}
