% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eur_map_g}
\alias{eur_map_g}
\title{Group labels for Europe map data}
\format{
A vector or data frame containing group labels.
}
\source{
Prepared for examples in the dppca package.
}
\usage{
data(eur_map_g)
}
\description{
Group labels corresponding to observations in \code{eur_map}.
}
\keyword{datasets}
