% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaussian_groups}
\alias{gaussian_groups}
\title{Gaussian grouped example data}
\format{
A data frame or matrix.
}
\source{
Simulated for examples in the dppca package.
}
\usage{
data(gaussian_groups)
}
\description{
A simulated Gaussian dataset with group structure.
}
\keyword{datasets}
