% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lambda.R
\name{lambda_mono_decreasing_nonneg}
\alias{lambda_mono_decreasing_nonneg}
\title{Enforce nonnegative nonincreasing sequence}
\usage{
lambda_mono_decreasing_nonneg(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A numeric vector of the same length.
}
\description{
Enforce nonnegative nonincreasing sequence
}
\keyword{internal}
