% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pca_helpers.R
\name{mech_tau_sph}
\alias{mech_tau_sph}
\title{Gaussian mechanism applied to spherical Kendall matrix}
\usage{
mech_tau_sph(X, sig, cpp.option = FALSE)
}
\arguments{
\item{X}{Numeric matrix.}

\item{sig}{Noise standard deviation.}

\item{cpp.option}{Logical; currently only FALSE is supported.}
}
\value{
A noisy symmetric d x d matrix.
}
\description{
Gaussian mechanism applied to spherical Kendall matrix
}
\keyword{internal}
