% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lambda.R
\name{orthonormalize_basis}
\alias{orthonormalize_basis}
\title{Re-orthonormalize basis vectors}
\usage{
orthonormalize_basis(V, k)
}
\arguments{
\item{V}{Numeric matrix.}

\item{k}{Number of columns to retain.}
}
\value{
A d x k orthonormal matrix.
}
\description{
Re-orthonormalize basis vectors
}
\keyword{internal}
