% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lambda.R
\name{prep_matrix_for_pca}
\alias{prep_matrix_for_pca}
\title{Preprocess matrix for PCA}
\usage{
prep_matrix_for_pca(X, center = TRUE, standardize = TRUE)
}
\arguments{
\item{X}{Numeric matrix-like object.}

\item{center}{Logical; whether to center columns.}

\item{standardize}{Logical; whether to scale columns by their SD.}
}
\value{
A numeric matrix.
}
\description{
Preprocess matrix for PCA
}
\keyword{internal}
