% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dp_app_data.R
\name{run_dp_app_data}
\alias{run_dp_app_data}
\title{Run DP PCA score-plot Shiny app for user-supplied data}
\usage{
run_dp_app_data(X, G = NULL, title = "DP PCA Score Plot (User data)")
}
\arguments{
\item{X}{Numeric matrix/data.frame (n x p).}

\item{G}{Optional group vector (length n) or a column name in X.}

\item{title}{App title.}
}
\description{
Run DP PCA score-plot Shiny app for user-supplied data
}
