% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pca_helpers.R
\name{tau_sph}
\alias{tau_sph}
\title{Sample multivariate Kendall's tau matrix under spherical transformation}
\usage{
tau_sph(X, cpp.option = FALSE)
}
\arguments{
\item{X}{Numeric matrix.}

\item{cpp.option}{Logical; currently only FALSE is supported.}
}
\value{
A symmetric d x d matrix.
}
\description{
Sample multivariate Kendall's tau matrix under spherical transformation
}
\keyword{internal}
