% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_lambda.R
\name{validate_dp_pca_inputs}
\alias{validate_dp_pca_inputs}
\title{Validate common PCA inputs}
\usage{
validate_dp_pca_inputs(X, k, eps_total, delta_total)
}
\arguments{
\item{X}{Numeric matrix-like object.}

\item{k}{Number of components.}

\item{eps_total}{Privacy epsilon.}

\item{delta_total}{Privacy delta.}
}
\value{
Invisibly returns TRUE.
}
\description{
Validate common PCA inputs
}
\keyword{internal}
