% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult data example}
\format{
A data frame or matrix.
}
\source{
Prepared for examples in the dppca package.
}
\usage{
data(adult)
}
\description{
A sample dataset used for examples in the dppca package.
}
\keyword{datasets}
