% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{adult}
\alias{adult}
\title{Numeric subset of Adult dataset}
\format{
A data frame with numeric variables:
\describe{
\item{age}{age}
\item{education_num}{years of education}
\item{capital_gain}{capital gains}
\item{capital_loss}{capital losses}
\item{hours_per_week}{hours worked per week}
}
}
\usage{
data(adult)
}
\description{
Numeric subset of Adult dataset
}
\examples{
data(adult)
summary(adult)
}
\keyword{datasets}
