% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scree_controls.R
\name{clipped_control}
\alias{clipped_control}
\title{Control options for clipped scree estimation}
\usage{
clipped_control(C_clip)
}
\arguments{
\item{C_clip}{Positive clipping threshold for squared centered principal
component scores. This value has no default because it should be chosen
according to the scale of the data.}
}
\value{
A list of control options for \code{method = "clipped"}.
}
\description{
Creates a control list for the clipped-mean scree estimator used by
\code{\link[=dp_scree]{dp_scree()}} and \code{\link[=dp_scree_plot]{dp_scree_plot()}} when \code{method = "clipped"}.
}
\details{
The clipped method estimates each scree value by clipping the squared scores
at \code{C_clip} and then applying a sensitivity-calibrated Gaussian mechanism
\insertCite{dwork2014algorithmic}{dppca}. Larger values of \code{C_clip} reduce
clipping bias but increase the sensitivity, and therefore the scale of
the privacy noise.
}
\examples{
clipped_control(C_clip = 3)

}
\references{
\insertRef{dwork2014algorithmic}{dppca}
}
\seealso{
\code{\link[=dp_scree]{dp_scree()}} for using clipped scree estimation.
\code{\link[=dp_scree_plot]{dp_scree_plot()}} for plotting private scree estimates.
}
