% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dppca-package.R
\docType{package}
\name{dppca-package}
\alias{dppca}
\alias{dppca-package}
\title{dppca: Differentially Private PCA Scree and Score Visualization}
\description{
Tools for performing principal component analysis and constructing
differentially private scree estimates and score-based visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yejinjo0220/dppca}
  \item \url{https://yejinjo0220.github.io/dppca/}
  \item Report bugs at \url{https://github.com/yejinjo0220/dppca/issues}
}

}
\author{
\strong{Maintainer}: Yejin Jo \email{yejinjo0220@gmail.com}

Authors:
\itemize{
  \item Minwoo Kim \email{mwkim.stat@gmail.com}
}

}
\keyword{internal}
