% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eur_map}
\alias{eur_map}
\title{European map genotype dataset}
\format{
A data frame with 1387 rows and 21 variables:
\describe{
\item{X.1, ..., X.20}{Numeric PCA scores for 20 principal components.}
\item{color}{Group labels.}
}
}
\usage{
data(eur_map)
}
\description{
European map genotype dataset
}
\examples{
data(eur_map)
str(eur_map)
}
\keyword{datasets}
