% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eur_map}
\alias{eur_map}
\title{European data (variables only)}
\format{
A data frame with 1387 rows and 20 numeric variables.
}
\usage{
data(eur_map)
}
\description{
European data (variables only)
}
\keyword{datasets}
