% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eur_map}
\alias{eur_map}
\title{European map genotype dataset}
\format{
A data frame with:
\describe{
\item{X}{A numeric matrix/data frame for PCA.}
\item{color}{Group labels extracted from POPRES dataset.}
}
}
\usage{
data(eur_map)
}
\description{
European map genotype dataset
}
\examples{
data(eur_map)
str(eur_map)
}
\keyword{datasets}
