% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{eur_map_g}
\alias{eur_map_g}
\title{European data with group labels}
\format{
A data frame with 1387 rows, 20 numeric variables, and 1 factor column (color).
}
\usage{
data(eur_map_g)
}
\description{
European data with group labels
}
\keyword{datasets}
