% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gau}
\alias{gau}
\title{Gaussian example data}
\format{
A data frame or matrix.
}
\source{
Simulated for examples in the dppca package.
}
\usage{
data(gau)
}
\description{
A simulated Gaussian dataset used for examples.
}
\keyword{datasets}
