% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gau}
\alias{gau}
\title{Gaussian mixture simulation (variables only)}
\format{
A data frame with 5000 rows and 20 numeric variables.
}
\usage{
data(gau)
}
\description{
Gaussian mixture simulation (variables only)
}
\keyword{datasets}
