% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gau_g}
\alias{gau_g}
\title{Group labeled Gaussian example data}
\format{
A vector or data frame containing group labels.
}
\source{
Simulated for examples in the dppca package.
}
\usage{
data(gau_g)
}
\description{
Gaussian example data \code{gau} with group labels.
}
\keyword{datasets}
