% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gau_g}
\alias{gau_g}
\title{Gaussian mixture simulation with group labels}
\format{
A data frame with 5000 rows, 20 numeric variables, and 1 factor column (color).
}
\usage{
data(gau_g)
}
\description{
Gaussian mixture simulation with group labels
}
\keyword{datasets}
