% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{gaussian_groups}
\alias{gaussian_groups}
\title{Gaussian mixture synthetic dataset}
\format{
A data frame with:
\describe{
\item{X}{A numeric matrix (5000 × 20).}
\item{color}{Group labels ("red", "orange", "green", "blue", "purple").}
}
}
\usage{
data(gaussian_groups)
}
\description{
Gaussian mixture synthetic dataset
}
\examples{
data(gaussian_groups)
table(gaussian_groups$color)
}
\keyword{datasets}
