% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_pca_helpers.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector to unit norm}
\usage{
normalize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector with Euclidean norm 1.
}
\description{
Normalize a vector to unit norm
}
\keyword{internal}
